Train display add, part 3. Train delay and cancellation added, ticket total price computation added.

+++ ticket-to-ride/src/main/java/org/tsystems/javaschool/model/dto/ticket/TicketDto.java
@@ -19,7 +19,7 @@ public class TicketDto implements Serializable {
 
     private int id;
     private PassengerDto passengerDto;
-    private int totalPrice;
+    private double totalPrice;
     private List<TicketScheduleSectionDto> ticketScheduleSectionDtoList;
 
     public TicketScheduleSectionDto getFirstSection() {
